feature: Implement the functionality of allowing an agent to be able to delete a conversation message or not

+++ app/controllers/api/v1/accounts/conversations/messages_controller.rb
@@ -12,6 +12,8 @@ class Api::V1::Accounts::Conversations::MessagesController < Api::V1::Accounts::
   end
 
   def destroy
+    return head :bad_request unless message.can_delete_message?
+
     ActiveRecord::Base.transaction do
       message.update!(content: I18n.t('conversations.messages.deleted'), content_attributes: { deleted: true })
       message.attachments.destroy_all
